Stops pretending that we can dynamically control the UI.
A UI framework should never try to outsmart the user. Forcing font sizes
and widget width is only going to annoy people and can easily be
circumvented. The proper way to ensure that apps follow certain constraints
is by providing:

 1. Good documentation
 2. Good error messages
 3. Quality assurance

Changing things at runtime like we were doing before this patch simply
is a solution looking for a problem.